Correct the computation for the end of a Block
The disassemble code was omitting the last instruction of each basic
block.  The Block class last method provides the address of the last
instruction of the basic block.  What really is wanted is the value
from the end method. It provides the address just past the end of the
basic block.