Add a CFGraph and disassemble examples from the ParseAPI manual (#19)
* Add disassemble and CFGraph examples the ParseAPI Programmer's Guide

* Fix disassemble example to compile

The InstructionDecoder class decode method returns class Instruction
not Instruction::Ptr. The code has been corrected to expect the
appropriate type from the decode method.

* Explicitly pass addresses to decode

The instruction decoder was set up to just decode a particular instruction
rather than a buffer the size of the function.  For this to
work need to explicitly include the address of the instruction to decode.

* Correct the computation for the end of a Block

The disassemble code was omitting the last instruction of each basic
block.  The Block class last method provides the address of the last
instruction of the basic block.  What really is wanted is the value
from the end method. It provides the address just past the end of the
basic block.