Add unusedArg example (#23)
The unusedArg example analyzes the entry to each function in an
x86_64, PowerPC, or AArch64 executable and lists out which arguments
are unused by the function.  This code takes a number of
simplifications by assuming the arguments are always simple integer
arguments in registers and never passed as floating point values.  It
also doesn't handle cases where functions have more arguments passed
via the stack.  In the case of x86_64 that is a maximum of six
arguments.