Check to see if symbol information found for the function
If there is no symbol information for the function then the code
cannot do the analysis and needs to move on to the next function.  If
findFuncByEntryOffset doesn't find associated symbol table
information, the method does not initialize the variable func_sym.
Thus, need to chekc whether it was successful before attempting to get
the parameters.